+++ jest.config.js
@@ -1,0 +1,16 @@
+module.exports = {
+  testEnvironment: 'node',
+  roots: [
+    '<rootDir>/packages',
+    '<rootDir>/example'
+  ],
+  transform: {
+    '^.+\\.tsx?$': 'ts-jest'
+  },
+  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json'],
+  coverageReporters: [
+    'json-summary',
+    'text',
+    'lcov'
+  ]
+}
