+++ packages/yk-cli/types/init.d.ts
@@ -1,0 +1,3 @@
+import { Optional } from './interface/option';
+declare const init: (option: Optional) => Promise<void>;
+export default init;
