+++ tsconfig.json
@@ -1,0 +1,20 @@
+{
+    "compilerOptions": {
+      "target": "esnext",
+      "module": "commonjs",
+      "lib": [
+        "esnext"
+      ],
+      "declaration": true,
+      "jsx": "react",
+      "declarationDir": "./types/",
+      "sourceMap": true,
+      "strict": true,
+      "moduleResolution": "node",
+      "esModuleInterop": true,
+      "rootDir": "./src",
+      "experimentalDecorators": true,
+      "resolveJsonModule": true,
+      "emitDecoratorMetadata": true /* Enables experimental support for emitting type metadata for decorators. */
+    }
+}
